% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Check_Keyword}
\alias{Check_Keyword}
\title{Checks if classification exists for a keyword}
\usage{
Check_Keyword(text, checking.df)
}
\arguments{
\item{text}{character vector - contains text corpus to be searched}

\item{checking.df}{dataframe - containing json keyword config information}
}
\value{
list of lists containing classifications found
}
\description{
Check if any classification can be applied to the text corpus for a given keyword and its context descriptions. Meant to be used internally by other functions.
}
\examples{
Check_Keyword(text_corpus, json_dataframe)
}
