% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Classify_Text}
\alias{Classify_Text}
\title{Perform the classification of the text corpuses}
\usage{
Classify_Text(text.df, column.name, json.file.path)
}
\arguments{
\item{text.df}{dataframe - dataframe containing the text corpuses}

\item{column.name}{string - name of the column containing the text corpus within the provided dataframe}

\item{json.file.path}{string - full file path of the json file}
}
\value{
dataframe provided in the first argument with columns CLASSIFICATION and CLASSIFICATION_REASON added
}
\description{
Perform the classification of the text corpuses
}
\examples{
Classify_Text(text.corpus.dataframe, "CHARTED_VALUE", "../keywords.json")
}
