% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Create_Perturbation_Freq_List}
\alias{Create_Perturbation_Freq_List}
\title{Create frequency list for keyword perturbation in text corpus}
\usage{
Create_Perturbation_Freq_List(textblock.array, word, distance)
}
\arguments{
\item{textblock.array}{character vector - containing list of texts to be searched}

\item{word}{string - keyword to be looked for}

\item{distance}{numerical - sensitivity of the fuzzy text matching (see agrep argument max.distance)}
}
\value{
dataframe containing keyword perturbations and its frequency
}
\description{
Create frequency list fo all different representations/mispellings/perturbations that exist in an array or single text block
}
\examples{
Create_Perturbation_Freq_List(text.corpus.dataframe$column.of.interest, "chest pain", 0.1)
}
