% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Create_Wordcloud}
\alias{Create_Wordcloud}
\title{Plot a word cloud}
\usage{
Create_Wordcloud(words, freq, min.freq = 1, max.words = 200,
  random.order = FALSE, rot.per = 0.35, colors = brewer.pal(8,
  "Dark2"))
}
\arguments{
\item{words}{character vector - contains words to be plotted}

\item{freq}{character vector - contains word's frequency}

\item{min.freq}{integer - words with frequency below min.freq will not be plotted}

\item{max.words}{integer - maximum number of words to be plotted. least frequent terms dropped}

\item{random.order}{boolean - plot words in random order. If false, they will be plotted in decreasing frequency}

\item{rot.per}{numerical - proportion words with 90 degree rotation}

\item{colors}{RColorBrewer - color words from least to most frequent}
}
\description{
Plot a word cloud
}
\examples{
Create_Wordcloud(words = wordslist, freq = freqlist, min.freq = 1, max.words=200, random.order=FALSE, rot.per=0.35, colors=brewer.pal(8, "Dark2"))
Create_Wordcloud(words = wordslist, freq = freqlist)
}
