% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Get_Occurrences}
\alias{Get_Occurrences}
\title{Function that extracts all perturbations of a given word in a text corpus accounting for mispellings}
\usage{
Get_Occurrences(text, word, distance)
}
\arguments{
\item{text}{Character vector - contains text corpus to be searched}

\item{word}{string - keyword to be looked for}

\item{distance}{numerical - sensitivity of the fuzzy text matching (see agrep argument max.distance)}
}
\value{
character vector containing all perturbations of the keyword found in the text corpus
}
\description{
Function that extracts all perturbations of a given word in a text corpus accounting for mispellings
}
\examples{
Get_Occurences("Pt has heavy chest pain", "chest pain", 0.1)
}
