% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Set_KWIC_Prefix_Remove_Regex}
\alias{Set_KWIC_Prefix_Remove_Regex}
\title{Set regex string to cleanse prefix words extracted}
\usage{
Set_KWIC_Prefix_Remove_Regex(regex.string)
}
\arguments{
\item{regex.string}{string - regex expression to be used}
}
\description{
Set regex string to cleanse prefix words extracted
}
\examples{
Set_KWIC_Prefix_Remove_Regex(".*\\\\.")
}
