% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLACK.R
\name{Set_KWIC_Size}
\alias{Set_KWIC_Size}
\title{Set number of prefix and suffix context words to be analysed for keyword in context}
\usage{
Set_KWIC_Size(size)
}
\arguments{
\item{integer}{input integer value > 0}
}
\value{
numeric value 0 if successfully changed or -1 if not successful
}
\description{
Set number of prefix and suffix context words to be analysed for keyword in context
}
\examples{
Set_KWIC_Size(4)
}
\seealso{
\code{\link{kwic}} value set by this function is used by "window" parameter of the function
}
